core: add a simple unit test to start things out, fix bugs, refactor
The original code was utterly broken and this change fixes many of the
original problems.  Having devicecloud.devicecloud was leading to import
issues and this allows for a simpler interface from the library users
perspective.

The test itself just tests get_devices() but it still found a bug.  It looks
like httpretty works well for mocking out http requests (it monkey patches
socket calls).