devicecore: rework how devicecore construction happens and greatly expand
Previously, the factory method would parse selected parts of the json
and then pass all of these items into the Device object.  Now we pass
in the cached data we have but compute as much as possible and in a way
that allows for "fresh" data to be retrieved (update cache) at any point
along the way.

Some initial work has been done to pave the way for additional functionality
by passing a reference to the SCI API to Device instances.