refactor: remove extra "api" namespace that provides no value
Since we now require that users import things out of modules like
"streams", we should remove the "api" namespace which is pretty
meaningless as basically the only we expose is the "api".

Zen of Python: "Flat is better than nested"

This change also inadvertedly includes some work in progress that is being
done in order to expose a better API for data points (untested).