PYTHONDC-8: Tox testing support and port to Python3.3+
This commit adds support for python 3 (with the help of six) while
keeping backwards compatability with python 2.7.  Currently, python 3.2
has a single failure (which is caught by Tox!)

It is assumed that tox will be run with pythonbrew in order to magically
acquire all the required versions of the python interpreter for testing.
A script has been provided which takes care of those details.