PYTHONDC-12: Use pyenv rather than pythonbrew for building interpreters
This works in a very similar way to how things worked with pythonbrew
but using pyenv which is actively being supported.  As a nice side
effect of the switch, we are now able to define a single tox.ini
that we should also be able to use with a service like travisci.