streams: ensure stream_id is stripped of leading slashes and fix problem with string_types
Previously, there could have been confusion with leading slashes being
included in stream_ids.  We still accept these if given as such, but we
will always strip the id internally so that we have some consistency.

I debated whether we should include or exclude the slash, but responses
from the device cloud do _not_ have the slash, so I am using that as
the canonical form of a streamId.