PYTHONDC-6: Defining new filedata api and API for building conditions
This commit begins to rework the filedata API and tries to simplify
one of the more complex tasks required for interacting with the device
cloud filedata API, which is building 'condition' query parameters for
filtering which files one should see.  The approach here seems
promising and is inspired by the syntax provided by ORM tools like
sqlalchemy and django.