PYTHONDC-43: Add support for deleting data points
This commit adds basic support for deleting individual data points
and for deleting data points within a provided time window from
a stream.

During the course of implementing this change, it became
clear that there was also a need to rework how datetime objects were
handled within the library.  All datetime objects should now be
converted to be UTC when coming into the library or entering it
from the device cloud.  The arrow library is used to facilitate
these transformations.