docs: fix sphinx warnings and expand on docs for get_devices()
This started as just addressing warnings, but I decided to add some
examples and further clarification to get_devices() while I was making
changes.  This is likely to be one of the methods that almost every
developer will use and I thought the addition of some additional
examples was justifiable.