core: improve documentation and provide better error messages for HTTP
Previously, when a call failed (for any unhandled reason), the user
of the library would just get an exception indicating that a
DeviceCloudHttpException occurred without much additional information.

This documents the 'response' attribute (now a property) on the exception
object and adds a __str__ method that will print the failing response
body on failure.

This is related to PYTHONDC-96.

Signed-off-by: Paul Osborne <paul.osborne@etherios.com>