core: do not encode url path prior to hitting requests library
This fixes an issue introduced somewhere in the past few versions
of requests (which we had not frozen previously).  Rather than freezing
to an old version of requests, we bump the version to the latest and
make the required changes in order to make the libary work with requests.

This resolves PYTHONDC-97 against the issue.

Signed-off-by: Paul Osborne <osbpau@gmail.com>