test: streams: introduce infrastructure for integration tests
This commit introduces infrastructure for performing integration tests
against the device cloud and a few basic tests (based off of examples)
for streams.

There are still some kinks to work out but the basic concept appears
to work well.  Most of the problems being seen have to do with
throttling and not having a great way to "reset" the state of the
device cloud (particularly when things fail due to throttling).

Signed-off-by: Paul Osborne <paul.osborne@etherios.com>