core: add support for retry on HTTP throttling response
This changes resolves PYTHONDC-98 and introduces a solution for a
problem often encountered by users of the device cloud.  The problem
is that the device cloud has limits on the number of requests that
can be issued in each 10 second window, depending on one's account type.

Writing one's code to be knowledgeable of this limit can be a challenge
and presents a difficulty for users of the device cloud that they should
not have to be worried about.  This change should make it possible for
users of the python-devicecloud library to remain mostly ignorant of the
fact that throttling is taking place.  There will still, of course, be
an added delay on any web service call that happens to occur.  I am not
aware of a better alternative than is implemented here for avoiding the
problem.

Signed-off-by: Paul Osborne <paul.osborne@etherios.com>

Conflicts:
	devicecloud/__init__.py