ws: provide API for making authenticated web service requests
This commit adds a new API endpoint, ws, which provides some
syntactic sugar for getting nice looking paths to various
web service endpoints on the device cloud.

In addition, some textual documentation has been added that
covers this new interface and makes additional reference to
the DeviceCloudConnection APIs that are now part of the
public API for the library (previously private).

This work is based on the initial patch by
Dan Harrison <dan.harrison@digi.com>.

Signed-off-by: Paul Osborne <paul.osborne@etherios.com>