devicecore: provisioning: revised API/implementation for provisioning
This change adds documentation for provisioning and iterates on the
work done by Dan Harrison to create an API for provisioning that
is more consistent with the other APIs and which more fully exposes
the various available fields that can be provided when provisioning
a device in the cloud.

This changes relates to PYTHONDC-21 for single-device registration
and PYTHONDC-23 for registration of multiple devices.

Signed-off-by: Paul Osborne <paul.osborne@etherios.com>