build: fix failing (hanging) unit tests with python 3.4.1
There was a bug in version 3.4.0 and 3.4.1 of the interpreter that
caused the build to hang in an infinite loop.  The fix is to use
python 3.4.3 that appears to have a fix.  I have not determined the
exact change in the interpreter that resolves the issue.

Pyenv needs to be updaded in order to know how to build 3.4.3, so code
has been added that will ensure that pyenv gets updated each time
toxtests.sh is run.   The assumption is that the master branch for
pyenv will be reasonably stable.