monitor: deserialize json data on reception
This removes the burden of performing JSON unpacking from the user.
This assumes you are using JSON which is probably a reasonable assumption
but not limited by the API at present.

Some logging was also modified to be more clear with regards to
the return value from callbacks.

Relates    to:
- Monitor Support: https://jira.digi.com/browse/PYTHONDC-54
- TCP Mon Support: https://jira.digi.com/browse/PYTHONDC-55
- Delete Monitors: https://jira.digi.com/browse/PYTHONDC-56

Signed-off-by: Paul Osborne <Paul.Osborne@digi.com>