monitor: test: unit/integration tests for monitor implementation
This commit adds unit and integration tests for monitors.  The code
as-is for the TCP push monitor is hard to unit test as-is, so
for now an integration test has been put in place to provide some
level of automated coverage for this code.  It does not test
edge cases well but it is something.

This commit also includes a fair number of bug fixes for python3
with the new TCP push monitor implementation which previously was
written to support python 2 only.

Relates to:
- Monitor Support: https://jira.digi.com/browse/PYTHONDC-54
- TCP Mon Support: https://jira.digi.com/browse/PYTHONDC-55
- Delete Monitors: https://jira.digi.com/browse/PYTHONDC-56

Signed-off-by: Paul Osborne <Paul.Osborne@digi.com>