Fix an issue with encoding and decoding data types between the devicecloud and python
Previously the data was decoded (from dc type to python type) every time
get_data was called on the DataPoint.  This was not correct as for some
data types it may not be possible to call the decode method on the already
decoded data.  Further, the encode method was never properly used when
creating the XML to push data to the device cloud.  This issue was masked
by the fact that for currently supported data types the default str method
used was sufficient.

Signed-off-by: Zach Varberg <Zachary.Varberg@Digi.com>